+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.5] - 2026-02-17
+
+### Added
+
+- `WlMonitorAction::SetPosition { name, x, y }` - Set a monitor's position in the global coordinate space
+- `SetPosition` variant in `ActionKind` for error reporting
+
 ## [0.1.4] - 2026-02-16
 
 ### Added